requirements.txt
opencv-python
python-dotenv
google-generativeai
chromadb
sentence-transformers
